Sprint 3. RF TestPage and implementing generics (#73)
