+++ src/test/java/pages/BaseSideMenu.java
@@ -4,7 +4,7 @@ import com.microsoft.playwright.Locator;
 import com.microsoft.playwright.Page;
 import com.microsoft.playwright.Playwright;
 
-abstract class SideMenuPage extends BaseLocator{
+abstract class BaseSideMenu extends BaseLocator{
 
     private final Locator homeButton = exactButton("Home");
     private final Locator flashcardsButton = exactButton("Flashcards");
@@ -14,7 +14,7 @@ abstract class SideMenuPage extends BaseLocator{
     private final Locator studyGuideButton = button("Study guide");
     private final Locator performanceButton = exactButton ("Performance");
 
-    protected SideMenuPage(Page page, Playwright playwright) {
+    protected BaseSideMenu(Page page, Playwright playwright) {
         super(page, playwright);
     }
