+++ src/test/java/pages/HomePage.java
@@ -11,7 +11,8 @@ import io.qameta.allure.Step;
 
 import static java.lang.Integer.parseInt;
 
-public class HomePage extends SideMenuPage {
+public class HomePage extends BaseSideMenu {
+
     private final Locator studyThisButton = button("Study This");
     private final Locator testsButton = exactButton("Tests");
     private final Locator homeButton = exactButton("Home");
@@ -21,8 +22,8 @@ public class HomePage extends SideMenuPage {
     private final Locator progressbarPoints = locator("div>svg.CircularProgressbar+div>span").first();
     private final Locator progressbarSideMenuPoints = locator("div:has(.CircularProgressbar)+span").first();
     private final Locator streaksButton = locator("button>svg+p").last();
-    private final Locator streaksModalWindow = locator("div[role='dialog']");
     private final List<Locator> listCheckboxes = checkBoxesAll("label:has(input)");
+    private final Locator streaksModalWindow = locator("div[role='dialog']");
     private final Locator checkboxImage = locator("label:has(input) svg");
 
     private final int checkBoxNumber = TestUtils.getRandomInt(0, listCheckboxes.size());
