+++ src/test/java/pages/StudyGuidePage.java
@@ -5,7 +5,8 @@ import com.microsoft.playwright.Page;
 import com.microsoft.playwright.Playwright;
 import utils.TestUtils;
 
-public class StudyGuidePage extends SideMenuPage {
+public class StudyGuidePage extends BaseSideMenu {
+
     private final Locator wordList = waitForListLoadedGetByText("Projections");
     private final Locator noteButton = button(getWordText());
     private final Locator noteTextAria = locator("//textarea");
