+++ src/test/java/tests/TestTutorTest.java
@@ -1,14 +1,19 @@
 package tests;
 
 import com.microsoft.playwright.Locator;
+import io.qameta.allure.Description;
+import io.qameta.allure.Story;
+import io.qameta.allure.TmsLink;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import pages.HomePage;
 import pages.PreconditionPage;
+import pages.TestListPage;
 import pages.TestTutorPage;
 import utils.TestData;
 
 import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
+import static utils.ProjectProperties.BASE_URL;
 
 public class TestTutorTest extends BaseTest {
 
@@ -55,6 +60,76 @@ public class TestTutorTest extends BaseTest {
 
         assertThat(testTutorPage.getReportAProblemModal()).isVisible();
         assertThat(testTutorPage.getReportSentSuccessfullyMessage()).hasText(TestData.REPORT_MESSAGE);
+    }
+
+    @Test(description = "TC1344-01 The single non-active Checkbox can be checked.")
+    @Description("Objective: To verify that a non-active checkbox ca be successfully checked.")
+    @Story("Tests")
+    @TmsLink("9lf328qwx4bv")
+    public void testTheSingleCheckboxCanBeChecked() {
+        String checkboxText = new HomePage(getPage(), getPlaywright())
+                .clickTestsMenu()
+                .cancelDialogIfVisible()
+                .clickDomainsButton()
+                .clickRandomCheckboxAndReturnItsName();
+
+        assertThat(new TestListPage(getPage(), getPlaywright()).checkIcon(checkboxText)).isVisible();
+    }
+
+    @Test(description = "TC1344-02 Execute Tutor Mode with a randomly selected checkbox in the Domain section.")
+    @Description("Objective: To verify that User can successfully activate Tutor mode by checking a random checkbox " +
+            "in the Domain section and entering valid data in the ‘Number of Questions’ field.")
+    @Story("Tests")
+    @TmsLink("9a0yelvj0jdh")
+    public void testTutorModeWithRandomCheckboxInDomain() {
+        TestTutorPage testTutorPage = new HomePage(getPage(), getPlaywright())
+                .clickTestsMenu()
+                .cancelDialogIfVisible()
+                .clickDomainsButton()
+                .clickRandomCheckbox()
+                .clickTutorButton()
+                .inputNumberOfQuestions(TestData.ONE)
+                .clickGenerateAndStartButton();
+
+        waitForPageLoad(TestData.TEST_TUTOR_END_POINT);
+
+        assertThat(getPage()).hasURL(BASE_URL + TestData.TEST_TUTOR_END_POINT);
+        assertThat(testTutorPage.getTestQuestion()).containsText(TestData.QUESTION_MARK);
+        Assert.assertTrue(testTutorPage.countTestRadioButtons() >= 1);
+    }
+
+    @Test
+    public void testTutorModeWithRandomCheckboxInChapter() {
+        TestTutorPage testTutorPage = new HomePage(getPage(), getPlaywright())
+                .clickTestsMenu()
+                .cancelDialogIfVisible()
+                .clickChaptersButton()
+                .clickRandomCheckbox()
+                .clickTutorButton()
+                .inputNumberOfQuestions(TestData.ONE)
+                .clickGenerateAndStartButton();
+
+        waitForPageLoad(TestData.TEST_TUTOR_END_POINT);
+
+        assertThat(getPage()).hasURL(BASE_URL + TestData.TEST_TUTOR_END_POINT);
+        assertThat(testTutorPage.getTestQuestion()).containsText(TestData.QUESTION_MARK);
+        Assert.assertTrue(testTutorPage.countTestRadioButtons() >= 1);
+    }
+
+    @Test
+    public void testAfterMarkingTheCardTheNumberOfMarkedCardsIncreasedBy1() {
+        new PreconditionPage(getPage(), getPlaywright())
+                .startTest(TestData.ONE);
+
+        Locator numberMarked = new TestTutorPage(getPage(), getPlaywright())
+                .clickMarkForReviewButton()
+                .clickEndButton()
+                .clickYesButton()
+                .clickSkipButton()
+                .clickCloseTheTestButton()
+                .getNumberMarked();
+
+        assertThat(numberMarked).isVisible();
     }
 
     @Test
